requirements.txt
<<<<<<< HEAD
=======
anyio==4.9.0
APScheduler==3.6.3
>>>>>>> bbcaf15 (Instagram bot)
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
<<<<<<< HEAD
=======
cachetools==4.2.2
>>>>>>> bbcaf15 (Instagram bot)
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
h11==0.14.0
<<<<<<< HEAD
=======
httpcore==1.0.7
httpx==0.28.1
>>>>>>> bbcaf15 (Instagram bot)
idna==3.10
msgpack==1.1.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
<<<<<<< HEAD
=======
python-environ==0.4.54
python-telegram-bot==22.0
pytz==2025.1
>>>>>>> bbcaf15 (Instagram bot)
redis==5.2.1
requests==2.32.3
selenium==4.29.0
setuptools==76.1.0
<<<<<<< HEAD
=======
six==1.17.0
>>>>>>> bbcaf15 (Instagram bot)
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
<<<<<<< HEAD
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
=======
tornado==6.4.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==1.26.15
>>>>>>> bbcaf15 (Instagram bot)
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
